spice-util: support for non-default GMainContext
When spice-gtk is used in an application with its own GMainContext, the
wrong context will be used leading to various issues.

https://developer-old.gnome.org/programming-guidelines/stable/main-contexts.html.en

We add a new API spice_util_set_main_context() which allows the caller
to pass in the main context for a main loop that the caller controls and
is responsible for.